+++ worlds/apquest/apquest/graphics/full_heart.png
