+++ worlds/apquest/apquest/graphics/half_heart.png
